requirements.txt
polars>=1.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
jupyterlab>=4.0.0
pytest>=8.0.0
pandera[polars]>=0.18.0
